Add vacation alias separated with a comma
The current implementation of the vacation tool adds the autoreply address to the list of alias without a preceding comma. For example, if a user's aliases field was `user@domain.com`, then it would have become `user@domain.comuser#domain.com@autoreply-domain.com`.

Furthermore, if the vacation autoreply was set to inactive, the autoreply address was not removed from the list of aliases.

This PR proposes a small fix for both issues. One rare issue still persists, which would require a more extensive rewrite of the existing code: If a user has no aliases at all (because they set "Keep e-mails" in the forward tool to false and did not configure any forward addresses), the autoreply address should actually be appended without the preceding comma.